requirements.txt
selenium~=4.3.0
pytest~=7.0.1
allure-pytest
